[BUGFIX] fix Cannot read property 'document' of undefined
When plotly.js is loaded in Chrome by the native ES6 import,
the loading crashes with Cannot read property 'document' of
undefined

The problem is that the bundler assumes that when a function
is executed without any this, the this will automatically be
set to window. This is not the case with ES6 modules, where
this will be undefined.